fix: add missing organization folder to configs folder of deployment template (#64)
The bootstrapped deployment would fail without an organization folder
when parsing the configs. Added a `github_owner_name` input so that this
folder will be created on render as
`{{.subdirectory}}/configs/{{.github_owner_name}}`.

The README previously broke the deployment because GHTM expects only
`.yaml` files in the configs directory. The README will now be rendered
in deployment's root folder instead.