feat: updated minty action to expire token in post processing (#159)
Moved the call to the minty server and into a new javascript action
called `minty`. I then added a post section that calls the GitHub API to
revoke the token in the cleanup phase.

I will follow up with changes to the `mint-token` action to call `minty`
once the `minty` action gets on main.