fix: treat a missing org level repo as a 'not found' scenario instead of a hard failure (#206)
Fixes #205 

Attempts to mint a token for a nonexistent repository were causing hard
failures instead of being handled correctly. In this scenario the
correct thing to do is to treat it as if the file is not found and
continue processing up the tree looking for other files that might
contain the requested scope.