feat: take the target org as a parameter on the request (#219)
This will default to the org in the OIDC claim.

Currently, requests use the parsed org name from the OIDC claim which is
fine if you are only ever requesting permissions to act on your own org,
but in the case of GitHub administration we need to be able to target an
org (A) from a central repository in another org (B).