feat: allow a config doc with no defined permissions that allows anything that the installed app allows (#228)
The current implementation requires the config file to call out which
specific sets of permissions are needed for a given scope. This is fine
in most cases where we want to request a lesser set of permissions than
the app provides, but in some cases we really just want to take whatever
the app has defined.

Adding a new permission becomes a multi-step process where we first
change the permissions on the app, then update every config file that
might need to use those permissions. This is particularly painful when
dealing with the `minty.yaml` stored in the `.minty` repository as we
more broadly rollout cross org automation of GitHub via terraform.