fix: requests with no permissions against a scope that supports installed permissions instead of a fixed list fails (#229)
Minty allows requests for without permissions in the request object.
This is uses the set of permissions defined at the scope level and
replaces the empty set in the request with what is in the scope.

Unfortunately, when a request like this asks for a scope that specifies
the special indicator "*", that says to use whatever the installed app
allows, the request fails because it has replaced the request
permissions (nil) with a value of {"*":"*"} which GitHub doesn't
understand.

The fix is to catch this scenario, and set the request permissions to
`nil` which indicates to GitHub to grant whatever permissions are
defined on the app.

This PR comes with additional testing logic around this area to
hopefully help catch future regressions as the logic here is a bit
complicated.