fix: invalid format when requesting all permissions (#230)
This PR fixes an issue where a nil permissions map was sent to the
GitHub API when requesting a token, causing a 500 error from GitHub

The fix initializes the permissions map to an empty map and only
populates it if permissions are provided. This ensures that the request
sent to GitHub is always in a valid format.

Additionally, this PR adds logging for the token request to improve
observability.